+++ pages/organization/index.vue
@@ -11,9 +11,6 @@
           <el-tab-pane label="Socila Profile" name="third">
             <SocialProfile v-if="activeName == 'third'" :user="user"/>
           </el-tab-pane>
-          <el-tab-pane label="Policies" name="fourth">
-            <Policies v-if="activeName == 'fourth'" :user="user"/>
-          </el-tab-pane>
         </el-tabs>
       </div>
       <div>
@@ -26,7 +23,6 @@
 <script setup>
 import Address from '~/pages/organization/Components/Address.vue';
 import Basic from '~/pages/organization/Components/Basic.vue';
-import Policies from '~/pages/organization/Components/Policies.vue';
 import SocialProfile from '~/pages/organization/Components/SocialProfile.vue';
 
 const activeName = ref("first");
